order_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0